Move error message formatting to new module
